define CAC_ADDR when CONFIG_BRCM_UPPER_768MB
For this, we just need to map CAC_ADDR(addr) to addr as the upper 768M
are all mapped as cached. Only CAC_ADDR is used by ALSA to
make sure the memory is indeed in kseg0. Any usage of UNCAC_ADDR
should not compile and will need to be addressed.

Change-Id: Ib7fbe4b71f56fd4f52cf3d41638ea571b82dbb53