ubi: reduce error messages when there are ECC errors.
This avoids flooding the log when the ubi detects lots of ECC errors (on
flakey or old flash).  Discovered this on an SLC NAND device that had been
stress tested.

The excess log messages could be printed to the serial console and stop the
whole system from running, which could lead to watchdog timeouts.

Change-Id: Id25dd52e5e4f8cca1e132bdba10aaa2b889d16dc