partitionmap: default to right partition map depending on flash type.
There are currently too many variations of CFE out there for us to trust
that the partitionver= variable will always be passed.  For example, our
current CFE doesn't remember to pass it automatically when you're using
nfsroot.  Also, if you install an old software version with an old CFE, it
will fail to pass in the right value.

Instead, let's use the same auto-detection logic as CFE does, so if
partitionver= isn't provided, we use version 1 or 2 as appropriate.

Removed the references to multiple NOR partition versions, because switching
versions didn't actually work anyway.  No sense pretending it does.

Changed the way the initial partition map is initialized (using a callback
from mtd) so that we can get a copy of the actual mtd object, which is what
tells us the size of the MTD device we're supposed to be partitioning.  I
would like to have the model number instead/also, but the mtd layer doesn't
seem to like to share that with anyone, so we'll have to content ourselves
with size.

While playing with this, I learned about Redboot partition tables, which in
retrospect is probably what we should have started using the moment we
realized we had more than one flash type.  Oh well.

Change-Id: If050630e1cda6bd19b39e53f7c53a26fcf47da69