Remove /proc/sys/dev/repartition/info, add nand_size_mb instead.
The info file was a bunch of unformatted text and wasn't used anywhere, and
wasn't in the spirit of a sysctl (which should just be a single number or
short string).  The information we really might want is the nand size, so
drop it in as its own int sysctl, using megabytes so it fits easily inside
an int.

While we're here, get rid of the unnecessary 'settings' struct, which can
just be static globals instead, which is more common for how the kernel does
things.

Change-Id: Idc3990f1d9915a7e86611d2ee0bc400174616335