HFB filter to prioritize multicast streams
This improves performance in the scenario described in b/7027057, still,
it doesn't work for more than 600Mbps.
Added the HFB filter to prioritize the multicast streams.
Also, I added two debugfs counters for the number of filtered packets.

Included the fix for the compilation breakage when CONFIG_NET_SCH_MULTIQ
was undefined (missing definitions for macros related to number of queues and
per queue buffer descriptors).

Worked with dmentz on this.

+ added subdirectories to the debugfs root directory for each interface,
this was the correct stats are shown

Change-Id: I3dfdd10330f1700ccb7f89d42ff241df18fe1f68