Broadcom drop MoCADriver_27_Sep_12_r31089
Copied kernel/bmoca.c and include/linux/bmoca.h from
MoCADriver_27_Sep_12_r31089
Remedy for b/6653592 i.e. Broadcom CSP #567480.

After complaining that the new kernel/bmoca.c makes a reference on line
1784 to clk_set_rate() which is undefined in our version of the kernel,
Broadcom's Chris Jaszczur responded on 09-Oct-2012:

"There should have been a patch to
power.c included in the MoCA release, but it didn't make it into the
file...

You can find the patch in the MoCA 2.6.2 release, or you can use the
previous version of bmoca.c/h with this release.  The changes in bmoca
aren't relevant for MoCA 1.1: a change to partially enable power save
states, which you aren't using, and the clk_set_rate change, which
merely ends up setting the clock to the default value that it was
already set to for MoCA 1.1 anyway."

Change-Id: I0d7e0dcbbc3fffbc09dff4c9643667c510193b3d