Move low priority rx from softirq to work queue
In a pristine kernel, NAPI processes incoming packets in the
NET_RX_SOFTIRQ. In the presence of high volumes of incoming traffic, one
CPU spends a significant amount of time executing this softirq which has
higher priority than any user space process. Hence, there is an imminent
danger of user space process starvation which used to cause degradation
of the TV experience (picture freezing, distortion due to packet loss).
To improve on this situation, we move the processing of incoming, low
priority packets out of NAPI and NET_RX_SOFTIRQ and into a work item
that is submitted to a dedicated work queue.  As opposed to softirqs,
work items run in process context and are therefore pre-emptable. We
define low priority packets as packets that are not crucial for the TV
application.

Change-Id: I88cedea81a52b70fc311dcf0f493423f25c1b9b6