Custom NMI handler panics, reboots on watchdog NMI
The standard NMI handler in the Linux kernel tries to kill the current
process and continue normal operation. This is not helpful because our
watchdog triggered NMI is usually not related to the current process.

Our custom NMI handler dumps the registers and calls panic() instead.

The NMI vector from Broadcom appears to be broken. It neither enables
coprocessor 0 nor does it set KSU to 0 (kernel privilege level). As a
result, the NMI handler ran in user privilege level and therefore did
not get very far. Let's fix this by enabling coprocessor 0 and setting
KSU to 0.

Change-Id: I5eeeefc420bc2ec1964f21a82027c14998e1b142