Revert "Move low priority rx from softirq to work queue"
This reverts commit e8bd6212a2d9f4f0a55dbbb7bfe8fcb3b762e3a2.

Removing this patch greatly reduces CPU load and packet loss when recording
several channels at once on my setup.  I suspect it may be the cause of
b/8195440 but in any case I think it needs further attention before we can
include it in a release.  Most likely there is a way to tweak it to not
drain so much CPU.

Change-Id: I61dbbfb5128ff96f3dfd136f0d493c0f755af827