Fix deadlock during soft reboot
brcm_machine_restart() calls smp_send_stop() which is bad because the
latter might try to grab spinlocks that are already held at this point.
This can result in deadlocks.

In CASE#622032, Broadcom recommended removing the call to
smp_send_stop() stating it is no longer necessary:

"This was added due to errata in 7400A0 (TP1 did not go back into a
known state across a chip master reset) about 6 years ago. This can be
removed now, since this h/w bug does not exist in 7425."

Change-Id: Iafc3b7c159b3ef108931fb9334526a85098c8828