NMI handler: restart right away instead of panic()
There are bug reports where an NMI was taken, but the box failed to
reboot. Instead, the hardware watchdog had to swing in to reset the box.
As a result, we lost the content of the dmesg ring buffer.

We suspect that panic() gets stuck somewhere. To avoid this, we call
machine_restart() right away instead of panic().

Change-Id: I53bb93ff25f938cd09c9c741f97e369fe9e64e98