Route watchdog triggered NMI to both CPU threads
The default setting of the Broadcom Mode Register (CP0 Register 22,
Select 1) is to route NMIs only to Thread 0. As a result, we only get
stacktraces from Thread 0.

With this change, the NMI handler prints a stacktrace before it changes
the NMI routing to route the already asserted NMI to Thread 1 as well.
This will immediately cause an NMI exception on Thread 1, and we get a
stacktrace from there, too.

Change-Id: Idc10ff5b17146c8c1edbc9e5548323627440e36a