Add per queue drop count to rx_over_errors.
rx_over_errors is not supposed to be zero while
eth0/bcmgenet_discard_cnt_q0 increments.

This partially fixes b/8394453.

Additionally, rx_fifo_errors counter is used when the dmaFlag has DMA_RX_OV
in it instead of rx_over_errors so we can differentiate betweeen fifo
errors and insufficient descriptors errors.

Note: I was't able to test the situation where DMA_RX_OV flag is present
but the risk is minimal in this case.

Change-Id: I5b673f0c95289b624ee0214fcb5411cd6b4fca03