Trigger the OOM killer when thrashing is detected.
A high frequency of try_to_free_pages(), continuing for several consecutive
seconds, means we're severely low on memory.  On a normal Linux system with
swap space, maybe we might hope to eventually recover.  But on a system with
no swap space, it just means we have processes fighting over whose code gets
to be in RAM at once, and it'll all end in tears eventually.  Moreover, if
we don't kill one of them soon to free up some memory, we could end up in an
even worse situation where *all* the tasks in the whole system are
thrashing, trapped in uninterruptible sleep waiting for their pages, but
there's no memory left to actually finish the request.  And the OOM killer
can't kill a task until its outstanding request finishes.  If we run the OOM
killer a bit sooner, we have a better chance of recovery.

Change-Id: I5c236d9c3d54dd29d6d45d31a2531933aae35bdf