Make the OOM killer more murderous.
In the normal kernel, the OOM killer will pick a target, send it SIGKILL,
and wait forever for that target to die.  Unfortunately, on our system, if
memory gets extremely tight, and particularly if the kernel ends up using a
lot of memory for non-reclaimable socket buffers, there may be no way to
free up enough memory to kill the task that originally caused the problem,
because it's stuck in uninterruptible wait.  Instead, we might have to kill
off other processes until we get lucky enough to break the deadlock.  On my
test system in a stress test running the 'balloon' program, this patch takes
things from dying in less than 60 seconds to lasting for at least half an
hour.  There's a lot of collateral damage (tasks getting killed off) in the
meantime, but we have babysitter processes with oom_score_adj=-500 that will
restart things, and it seems to work.

Specific changes:

- Revise my earlier change that would sleep for a short time when congested.
  Instead, only realtime processes will sleep (since they won't give up
  their timeslice otherwise, even with cond_resched()).  Other tasks can
  safely use cond_resched(), and they should, because if they sleep while
  the OOM killer's target is waiting on them to release a page (or finish
  retrieving a page that's shared with the OOM target, such as a page from
  libc), they might never finish.

- boost_dying_task_prio() would set the dying task to a realtime priority of
  1, to help it die faster than average tasks.  However, we have tasks that
  (perhaps unnecessarily) run at rt priorities up to 99. Our OOM killed task
  needs to be boosted to a level >= those tasks in order to be guaranteed to
  run soon enough to free up pages.

- If the OOM killer triggers while another task is dying, choose a
  different task rather than just re-killing the original one.  To prevent
  this from causing a total killing spree if the OOM killer is called
  several times per second - which does happen sometimes when the system is
  desperate - we rate limit the OOM killer to once every 2 seconds.  That
  should give each task time to die before we kill the next one.

- Rate limit the stack and memory traces from "page allocation failure"
  messages.  These happen a lot in low-memory systems, and they happen from
  interrupt handlers, so the system can die just because it's telling you
  about a minor problem.

- In do_try_to_free_pages(), tell wait_iff_congested to wait for 10ms
  instead of 100ms.  When we have a lot of tasks thrashing and trying to die
  at once, we run into problems where two tasks end up waiting on the same
  shared pages at the same time, where one task is responsible for loading
  the page, and the others end up waiting for it.  If it's sleeping, it
  can't load the page.  This kind of intense problem actually clears up
  faster if the system thrashes harder.

Change-Id: I45511c62c6d53746f2eceaea54434bc08be16224