Remove extra kernel debugging options.
These options are useful and help us track down problems,
but they slow the product to the point where it is mostly
unpleasant to use. They also increase the memory footprint
substantially, leading to OOM Kills under lighter load
than should happen.

In the future this commit can be reverted to restore the extra
debugging, and revert the revert to remove them again.

This reverts commit e6886e861e1eccc994914d541e7ca5845d1dc329.
This reverts commit dc6a13b883e90c30dcb072a443bbbd1d39320955.
This reverts commit 20ce919c4db140530fb87eb5cd378c8e27b61908.

Change-Id: I8b627b1bec15745d818388c6202e09dfc5d365c1